requirements.txt
torch==2.9.1
torchvision==0.24.1
hydra-core==1.3.2
omegaconf==2.3.0
rich
moviepy==1.0.3
imageio
ffmpeg
huggingface-hub[hf-transfer,cli]
gymnasium
flask
datasets==3.3.2
gdown
einops==0.8.1
diffusers==0.30.3
wandb==0.22.1
av
pymunk==6.11.1
zarr
robomimic
mujoco==3.3.0
dm_control==1.0.28
fvcore
torchcfm==1.0.5
timm==1.0.20
numpy==1.26.4
